move lists to client-list page component
